All the changes

+++ DreamMachine/src/assignment/Response.java
@@ -22,16 +22,18 @@ public class Response extends Answer {
 	/*
 	 * Function that returns a string object version of the 
 	 * response question. It returns a string that has
-	 * the correct answerOptions separated by commas for a particular field 
-	 * separated by deliminators. 
+	 * the correct SEPERATOR separated answer options for a particular field 
+	 * separated by DELIM. 
 	 */
+	
+	@Override
 	public String toString () {
 		String options = "";
 		for  (int listIndex = 0, listIndexLen = answerOptions.size(); listIndex < listIndexLen; listIndex++){
 			if (listIndex != 0) options += " " + DELIM;
 			ArrayList<String> answerList = answerOptions.get(listIndex);
 			for (int i = 0, len = answerList.size(); i < len; i++) {
-				if (i != 0) options += ", ";
+				if (i != 0) options += SEPERATOR;
 				options += answerList.get(i); 
 			}	 
 		}
@@ -45,13 +47,13 @@ public class Response extends Answer {
 	 * answers list.
 	 */
 	public void parseOptions (String options) {
-		//example of options string => "Obama, Barrack, Barrack Obama <> The White House, White House"
+		//example of options string => "Obama SEPERATOR Barrack SEPERATOR Barrack Obama DELIM The White House SEPERATOR White House"
 		
 		String [] parsedAnswerLists = options.split(DELIM);
 		
 		for (int i = 0, len = parsedAnswerLists.length; i < len; i++) {
 			String parsedAnswerListStr = parsedAnswerLists[i]; 
-			String [] parsedAnswerArray = parsedAnswerListStr.split(", ");
+			String [] parsedAnswerArray = parsedAnswerListStr.split(SEPERATOR);
 			answerOptions.add(new ArrayList<String>(Arrays.asList(parsedAnswerArray)));
 		}
 	}
@@ -78,24 +80,24 @@ public class Response extends Answer {
 				}
 			}
 		}
-		
 		for (boolean isCorrect : correctResponses) if (!isCorrect) return false; 
 		return true;
 	}
 	
 	/*
-	 * Function that takes in a string that is an answer
-	 * and adds it into the all answer list.
+	 * Function that takes in a string of valid answer options for
+	 * and answer and adds it into the answer options list
 	 */
 	public void addAnswer (String answer) {
-		String [] parsedAnswerArray = answer.split(", ");
+		String [] parsedAnswerArray = answer.split(SEPERATOR);
 		answerOptions.add(new ArrayList<String>(Arrays.asList(parsedAnswerArray)));
 		numAnswers++;
 	}
 	
 	/*
-	 * Function that takes in a string that is an answer
-	 * and adds it into the all answer list.
+	 * Function that takes in a string that is a valid option and the 
+	 * index of the answer it belongs to. Adds it as an option to the
+	 * appropriate answers option list. 
 	 */
 	public boolean addToAnswer (String answer, int index) {
 		if (answerOptions.size() < index || index < 0) return false; 
@@ -103,36 +105,34 @@ public class Response extends Answer {
 	}
 	
 	/*
-	 * Function that takes in the option that the user
-	 * wants to delete and removes it from the list of 
-	 * options. It also checks if the option was a selected 
-	 * answer. If so, it removes it from the answerOptions list
-	 * as well.
+	 * Function that takes in the index of the answer options 
+	 * that the user want to delete and removes it from the list of 
+	 * answer options.
 	 */
 	public boolean deleteAnswer (int index) {
 		if (answerOptions.size() < index || index < 0) return false; 
 		answerOptions.remove(index);
 		numAnswers--;
 		return true; 
-
 	}
 	
 	/*
-	 * Function that takes in a string answer and removes it 
-	 * from the list of answerOptions. 
+	 * Function that takes in the index of the answer options and the option 
+	 * that user wants removed. Removes it as an option to the
+	 * appropriate answers option list. 
 	 */
 	public boolean deleteFromAnswer (String answer, int index) {
 		if (answerOptions.size() < index || index < 0) return false; 
-		answerOptions.get(index).remove(answer);
+		if (!answerOptions.get(index).remove(answer)) return false;
 		if (answerOptions.get(index).isEmpty()) deleteAnswer(index);
 		return true;
 	}
 
 	/*
-	 * Function that validates a multiple choice question.
+	 * Function that validates a response question.
 	 * First it checks that the size of answerOptions and options do not equal zero.
-	 * Second it checks that each answer is a valid option.
 	 */
+	@Override
 	public boolean isValid () {
 		if (answerOptions.size() == 0 || numAnswers == 0) return false;
 		for (ArrayList<String> answerList: answerOptions) {
@@ -140,7 +140,6 @@ public class Response extends Answer {
 		}
 		return true;
 	}
-	
 	
 	public static void main (String[] args) {
 		Response response1 = new Response();
